pyproject.toml
[tool.poetry]
name = "bytescraper"
version = "0.1.0"
description = ""
authors = ["Luc Zbonack <lmzbonack@email.arizona.edu>"]

[tool.poetry.dependencies]
python = "^3.6"
beautifulsoup4 = "^4.8"
requests = "^2.22"
django = "^2.2"
djangorestframework = "^3.10"
djongo = "^1.2"
django-model-utils = "^3.2"

[tool.poetry.dev-dependencies]
pytest = "^3.0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
